File clean up
And fixed a few more code smells
