+++ src/lib/xml/meta_handling.py
@@ -1,0 +1,2 @@
+# TODO: Do.
+# All xml handling should live in this folder. All metadata typing / metatyping should live here.
